Put the controllers inside the modules api and v1 and organize them in api/v1/
